updated for 4 channels

+++ ipmbbridged.cpp
@@ -518,14 +518,13 @@ end:
 
 IpmbChannel::IpmbChannel(boost::asio::io_service &io,
                          uint8_t ipmbBmcSlaveAddress,
-                         uint8_t ipmbRqSlaveAddress, 
-                         int channelIdx, ipmbChannelType type,
+                         uint8_t ipmbRqSlaveAddress, uint8_t channelIdx,
+                         ipmbChannelType type,
                          std::shared_ptr<IpmbCommandFilter> commandFilter) :
     i2cSlaveDescriptor(io),
     ipmbBmcSlaveAddress(ipmbBmcSlaveAddress),
-    ipmbRqSlaveAddress(ipmbRqSlaveAddress), 
-    channelIdx(channelIdx),
-    type(type), commandFilter(commandFilter)
+    ipmbRqSlaveAddress(ipmbRqSlaveAddress), channelIdx(channelIdx), type(type),
+    commandFilter(commandFilter)
 {
 }
 
@@ -537,7 +536,7 @@ int IpmbChannel::ipmbChannelInit(const char *ipmbI2cSlave)
     std::string busStr = ipmbI2cSlaveStr.substr(findHyphen + 1);
     try
     {
-        ipmbBusId = std::stoi(busStr); 
+        ipmbBusId = std::stoi(busStr);
     }
     catch (std::invalid_argument)
     {
@@ -686,7 +685,7 @@ uint8_t IpmbChannel::getRqSlaveAddress()
     return ipmbRqSlaveAddress;
 }
 
-int IpmbChannel::getChannelIdx()
+uint8_t IpmbChannel::getChannelIdx()
 {
     return channelIdx;
 }
@@ -768,7 +767,7 @@ std::tuple<int, uint8_t, uint8_t, uint8_t, uint8_t, std::vector<uint8_t>>
     return returnStatus(ipmbResponseStatus::timeout);
 }
 
-static IpmbChannel *getChannel(int channelIdx)
+static IpmbChannel *getChannel(uint8_t channelIdx)
 {
     auto channel =
         std::find_if(ipmbChannels.begin(), ipmbChannels.end(),
@@ -799,7 +798,7 @@ static int initializeChannels()
     }
     try
     {
-	int channelIdx = 0;
+        uint8_t channelIdx = 0;
         auto data = nlohmann::json::parse(configFile, nullptr);
         for (const auto &channelConfig : data["channels"])
         {
@@ -808,17 +807,16 @@ static int initializeChannels()
             uint8_t bmcAddr = channelConfig["bmc-addr"];
             uint8_t reqAddr = channelConfig["remote-addr"];
             ipmbChannelType type = ipmbChannelTypeMap.at(typeConfig);
-
-            auto channel = ipmbChannels.emplace(ipmbChannels.end(), io, bmcAddr,
-                                                reqAddr, channelIdx, type, 
-                                                commandFilter); 
+            auto channel =
+                ipmbChannels.emplace(ipmbChannels.end(), io, bmcAddr, reqAddr,
+                                     channelIdx, type, commandFilter);
             if (channel->ipmbChannelInit(slavePath.c_str()) < 0)
             {
                 phosphor::logging::log<phosphor::logging::level::ERR>(
                     "initializeChannels: channel initialization failed");
                 return -1;
             }
-	    channelIdx++;
+            channelIdx++;
         }
     }
     catch (nlohmann::json::exception &e)
@@ -894,8 +892,8 @@ void addUpdateSlaveAddrHandler()
             message.read(reqChannel, busId, slaveAddr);
 
             IpmbChannel *channel = getChannel(reqChannel);
-            if (channel == nullptr || 
-                  ( channel->getChannelType() == ipmbChannelType::me) ) 
+            if (channel == nullptr ||
+                (channel->getChannelType() == ipmbChannelType::me))
             {
                 phosphor::logging::log<phosphor::logging::level::ERR>(
                     "addUpdateSlaveAddrHandler: invalid channel");
